backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
authlib==1.3.0
click==8.2.1
fastapi==0.116.1
google-api-python-client==2.120.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
h11==0.16.0
idna==3.10
langchain==0.2.1
numpy==2.3.2
openai==1.30.1
pgvector==0.4.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sentence-transformers==2.3.1
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
